baseicon brokenness?
