+++ pinegrow.json
@@ -30,25 +30,25 @@
     "src/App.vue": {
       "frameworks": [
         "vue-router",
-        "@vue/devtools",
-        "vite-plugin-vue-devtools",
-        "vite-plugin-vue-layouts",
-        "unplugin-vue-router",
-        "unplugin-auto-import",
-        "unplugin-vue-components",
-        "@unocss/preset-icons",
-        "@pinegrow/vite-plugin",
-        "@vitejs/plugin-vue",
-        "vue",
-        "app_designer",
-        "project",
         "pg.insight.events",
         "pg.vue.tree.painter",
+        "@vue/devtools",
         "pg.css.grid",
+        "vite-plugin-vue-devtools",
         "pg.image.overlay",
+        "vite-plugin-vue-layouts",
         "pg.code-validator",
         "pg.project.items",
+        "unplugin-vue-router",
         "pg.asset.manager",
+        "project",
+        "unplugin-auto-import",
+        "app_designer",
+        "unplugin-vue-components",
+        "vue",
+        "@vitejs/plugin-vue",
+        "@pinegrow/vite-plugin",
+        "@unocss/preset-icons",
         "tw.flowbite",
         "tw",
         "pg.html"
@@ -57,25 +57,25 @@
     "src/views/MainPage.vue": {
       "frameworks": [
         "vue-router",
+        "pg.insight.events",
         "@vue/devtools",
+        "pg.vue.tree.painter",
         "vite-plugin-vue-devtools",
         "vite-plugin-vue-layouts",
+        "pg.css.grid",
         "unplugin-vue-router",
+        "pg.image.overlay",
         "unplugin-auto-import",
+        "pg.code-validator",
         "unplugin-vue-components",
+        "project",
+        "pg.project.items",
         "@unocss/preset-icons",
+        "pg.asset.manager",
         "@pinegrow/vite-plugin",
         "@vitejs/plugin-vue",
         "vue",
         "app_designer",
-        "project",
-        "pg.insight.events",
-        "pg.vue.tree.painter",
-        "pg.css.grid",
-        "pg.image.overlay",
-        "pg.code-validator",
-        "pg.project.items",
-        "pg.asset.manager",
         "tw.flowbite",
         "tw",
         "pg.html"
@@ -84,25 +84,25 @@
     "src/components/BaseIcon.vue": {
       "frameworks": [
         "vue-router",
+        "pg.insight.events",
         "@vue/devtools",
+        "pg.vue.tree.painter",
         "vite-plugin-vue-devtools",
         "vite-plugin-vue-layouts",
+        "pg.css.grid",
         "unplugin-vue-router",
+        "pg.image.overlay",
         "unplugin-auto-import",
+        "pg.code-validator",
         "unplugin-vue-components",
+        "project",
+        "pg.project.items",
         "@unocss/preset-icons",
+        "pg.asset.manager",
         "@pinegrow/vite-plugin",
         "@vitejs/plugin-vue",
-        "vue",
         "app_designer",
-        "project",
-        "pg.insight.events",
-        "pg.vue.tree.painter",
-        "pg.css.grid",
-        "pg.image.overlay",
-        "pg.code-validator",
-        "pg.project.items",
-        "pg.asset.manager",
+        "vue",
         "tw.flowbite",
         "tw",
         "pg.html"
@@ -137,35 +137,29 @@
     }
   },
   "template_framework_id": "vue",
-  "open-pages": [
-    "src/App.vue",
-    "src/components/CounterButton.vue",
-    "src/components/BaseIcon.vue",
-    "src/views/MainPage.vue",
-    "src/components/IconInclusion.vue"
-  ],
+  "open-pages": [],
   "breakpoints": ["640px", "768px", "1024px", "1280px", "1536px"],
   "frameworks": [
     "vue-router",
+    "pg.insight.events",
     "@vue/devtools",
+    "pg.vue.tree.painter",
     "vite-plugin-vue-devtools",
     "vite-plugin-vue-layouts",
+    "pg.css.grid",
     "unplugin-vue-router",
+    "pg.image.overlay",
     "unplugin-auto-import",
+    "pg.code-validator",
     "unplugin-vue-components",
+    "project",
+    "pg.project.items",
     "@unocss/preset-icons",
+    "pg.asset.manager",
     "@pinegrow/vite-plugin",
     "@vitejs/plugin-vue",
     "vue",
     "app_designer",
-    "project",
-    "pg.insight.events",
-    "pg.vue.tree.painter",
-    "pg.css.grid",
-    "pg.image.overlay",
-    "pg.code-validator",
-    "pg.project.items",
-    "pg.asset.manager",
     "tw.flowbite",
     "tw",
     "pg.html"
